--- Source/Hosting/Hosting.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <AssemblyName>Dolittle.Hosting</AssemblyName>
-  </PropertyGroup>
-
-  <Import Project="../../Build/MSBuild/default.props"/>
-
-  <ItemGroup>
-    <ProjectReference Include="../Booting/Booting.csproj"/>
-    <ProjectReference Include="../Concepts/Concepts.csproj"/>
-    <ProjectReference Include="../Configuration/Configuration.csproj"/>
-    <ProjectReference Include="../Serialization.Json/Serialization.Json.csproj"/>
-    <ProjectReference Include="../Types/Types.csproj"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Google.Protobuf" Version="3.5.1" />
-    <PackageReference Include="Grpc" Version="2.23.0" />
-  </ItemGroup>
-</Project>
