--- Specifications/Hosting/Hosting.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <AssemblyName>Dolittle.Hosting.Specs</AssemblyName>
-  </PropertyGroup>
-
-  <Import Project="../../Build/MSBuild/specs.props"></Import>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\Source\Hosting\Hosting.csproj" />
-    <ProjectReference Include="..\..\Source\Types.Testing\Types.Testing.csproj" />
-  </ItemGroup>
-</Project>
