--- Specifications/Hosting/for_Hosts/given/ICanBindMyHostType.cs
@@ -1,12 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Dolittle. All rights reserved.
- *  Licensed under the MIT License. See LICENSE in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-namespace Dolittle.Hosting.given
-{
-    public interface ICanBindMyHostType : ICanBindServices 
-    {
-        
-    }
-}
