+++ Specifications/Services/for_Hosts/given/ICanBindMyHostType.cs
@@ -1,0 +1,12 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Dolittle. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+namespace Dolittle.Services.given
+{
+    public interface ICanBindMyHostType : ICanBindServices 
+    {
+        
+    }
+}
